lV-Authentication_6_Steps
lV-Authentication_6_Steps_to_Authorization_Mastery

+++ app/Http/Controllers/JobController.php
@@ -2,12 +2,14 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Models\Job;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class JobController extends Controller
 {
-    public function index() 
+    public function index()
     {
         $jobs = Job::with('employer')->latest()->simplePaginate(3);
 
@@ -16,58 +18,60 @@ class JobController extends Controller
         ]);
     }
 
-    public function create() 
+    public function create()
     {
         return view('jobs.create');
     }
 
-    public function show(Job $job) 
+    public function show(Job $job)
     {
-
         return view('jobs.show', ['job' => $job]);
     }
 
-    public function store() 
+    public function store()
     {
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required']
         ]);
-    
+
         Job::create([
             'title' => request('title'),
             'salary' => request('salary'),
             'employer_id' => 1
         ]);
-    
+
         return redirect('/jobs');
     }
 
-    public function edit(Job $job) 
+    public function edit(Job $job)
     {
         return view('jobs.edit', ['job' => $job]);
     }
 
-    public function update(Job $job) 
+    public function update(Job $job)
     {
+        Gate::authorize('edit-job', $job);
+
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required']
         ]);
-    
+
         $job->update([
             'title' => request('title'),
             'salary' => request('salary'),
         ]);
-    
+
         return redirect('/jobs/' . $job->id);
     }
 
-    public function destroy(Job $job) 
+    public function destroy(Job $job)
     {
-         // authorize (On hold...)
-    $job->delete();
+        Gate::authorize('edit-job', $job);
 
-    return redirect('/jobs');
+        $job->delete();
+
+        return redirect('/jobs');
     }
 }
